requirements.txt
jax==0.3.25
jaxlib==0.3.25
flax==0.6.2
optax==0.1.4
tokenizers==0.13.2
safetensors==0.2.5
absl_py==1.3.0
ml_collections==0.1.1
tqdm==4.64.1